requirements.txt
flask==2.1.1
requests==2.26.0
discord.py==2.0.1
